+++ qoffer-domain/src/main/groovy/life/qbic/portal/portlet/offers/search/SearchOffers.groovy
@@ -1,5 +1,8 @@
 package life.qbic.portal.portlet.offers.search
 
+import life.qbic.portal.portlet.SearchCriteria
+import life.qbic.portal.portlet.offers.OfferExporter
+
 /**
  * This use case searches for offers matching provided criteria.
  *
@@ -9,5 +12,23 @@ package life.qbic.portal.portlet.offers.search
  * @since: 1.0.0
  * @author: Tobias Koch
  */
-class SearchOffers {
+class SearchOffers implements SearchOffersInput{
+
+    private final OfferExporter exporter
+    private final SearchOffersOutput output
+
+    SearchOffers(SearchOffersOutput output, OfferExporter exporter){
+        this.output = output
+        this.exporter = exporter
+    }
+
+    @Override
+    void searchOffers(SearchCriteria searchCriteria) {
+
+    }
+
+    @Override
+    void downloadOfferAsPdf(String offerId) {
+
+    }
 }