+++ qoffer-infrastructure/src/main/groovy/life/qbic/portal/qoffer2/QOfferPortlet.groovy
@@ -25,7 +25,8 @@ import life.qbic.portal.qoffer2.web.StyledNotification
 @CompileStatic
 class QOfferPortlet extends QBiCPortletUI {
 
-    private DependencyManager dependencyManager;
+
+    private DependencyManager dependencyManager
 
     QOfferPortlet() {
         super()