+++ qoffer-infrastructure/src/main/groovy/life/qbic/portal/qoffer2/web/views/CreateCustomerView.groovy
@@ -1,18 +1,25 @@
-package life.qbic.portal.portlet
-
+package life.qbic.portal.qoffer2.web.views
 
 import com.vaadin.data.ValidationResult
 import com.vaadin.data.ValueContext
 import com.vaadin.data.validator.EmailValidator
 import com.vaadin.data.validator.StringLengthValidator
+import com.vaadin.server.Page
 import com.vaadin.server.UserError
 import com.vaadin.ui.Button
 import com.vaadin.ui.ComboBox
 import com.vaadin.ui.FormLayout
-import groovy.util.logging.Log4j2
-import life.qbic.portal.qoffer2.web.ViewModel
-import life.qbic.datamodel.persons.Affiliation
+import com.vaadin.ui.Notification
 import com.vaadin.ui.TextField
+
+import groovy.util.logging.Log4j2
+import life.qbic.datamodel.dtos.business.Affiliation
+import life.qbic.portal.qoffer2.web.StyledNotification
+import life.qbic.portal.qoffer2.web.ViewModel
+import life.qbic.portal.qoffer2.web.controllers.CreateCustomerController
+
+import java.beans.PropertyChangeEvent
+import java.beans.PropertyChangeListener
 
 /**
  * This class generates a Form Layout in which the user
@@ -21,35 +28,30 @@ import com.vaadin.ui.TextField
  * CreateCustomerView will be integrated into the qOffer 2.0 Portlet and provides an User Interface
  * with the intention of enabling a user the creation of a new Customer in the QBiC Database
  *
- * @since: 1.0.0*
+ * @since: 1.0.0
  */
 
 @Log4j2
 class CreateCustomerView extends FormLayout {
     final private ViewModel viewModel
+    final private CreateCustomerController controller
 
-    String firstName
-    String lastName
-    String email
-    Affiliation affiliation
-    HashMap customerInfo
-
+    TextField titleField
     TextField firstNameField
     TextField lastNameField
     TextField emailField
     ComboBox affiliationComboBox
     Button submitButton
 
-    private boolean firstNameValidity
-    private boolean lastNameValidity
-    private boolean emailValidity
-    private boolean affiliationValidity
+    StyledNotification failureNotification
+    StyledNotification successNotification
 
-    CreateCustomerView(ViewModel viewModel) {
+    CreateCustomerView(CreateCustomerController controller, ViewModel viewModel) {
         super()
+        this.controller = controller
         this.viewModel = viewModel
-        this.customerInfo = new HashMap()
         initLayout()
+        registerListeners()
     }
 
     /**
@@ -61,14 +63,17 @@ class CreateCustomerView extends FormLayout {
         //Generate FormLayout and the individual components
         FormLayout createCustomerForm = new FormLayout()
 
+        this.titleField = new TextField("Title")
+        titleField.setPlaceholder("customer title if any")
+
         this.firstNameField = new TextField("First Name")
-        firstNameField.setPlaceholder("First Name")
+        firstNameField.setPlaceholder("customer first name")
 
         this.lastNameField = new TextField("Last Name")
-        lastNameField.setPlaceholder("Last Name")
+        lastNameField.setPlaceholder("customer last name")
 
         this.emailField = new TextField("Email Address")
-        emailField.setPlaceholder("Email Address")
+        emailField.setPlaceholder("customer email address")
 
         generateAffiliationSelector(viewModel.affiliations)
         affiliationComboBox.emptySelectionAllowed = false
@@ -82,6 +87,9 @@ class CreateCustomerView extends FormLayout {
         createCustomerForm.addComponent(affiliationComboBox)
         createCustomerForm.addComponent(submitButton)
 
+
+        // Retrieve user input from fields and add them to the the Binder if entries are valid
+
         //Add Validators to the components
         this.addComponent(createCustomerForm)
     }
@@ -94,11 +102,75 @@ class CreateCustomerView extends FormLayout {
     private void generateAffiliationSelector(List<Affiliation> affiliationList) {
 
         this.affiliationComboBox =
-                new ComboBox<>("Select an Affiliation")
-        affiliationComboBox.setPlaceholder("Select Affiliation")
+                new ComboBox<>("Affiliation")
+        affiliationComboBox.setPlaceholder("select customer affiliation")
         affiliationComboBox.setItems(affiliationList)
-        affiliationComboBox.setItemCaptionGenerator({ Affiliation af -> af.groupName })
+        affiliationComboBox.setItemCaptionGenerator({ Affiliation affiliation -> affiliation.organisation })
         affiliationComboBox.setEmptySelectionAllowed(false)
     }
 
+    /**
+     * This is used to indicate whether all fields of this view are filled correctly.
+     * It relies on the separate fields for validation.
+     * @return
+     */
+    private boolean allValuesValid() {
+        return !firstNameField.getComponentError() \
+            && !lastNameField.getComponentError() \
+            && !emailField.getComponentError() \
+            && affiliationComboBox.getSelectedItem().isPresent()
+    }
+
+    void registerListeners() {
+        //Add Listeners to all Fields in the Formlayout
+        this.firstNameField.addValueChangeListener({ event ->
+            ValidationResult result = new StringLengthValidator(
+                    "Please input a valid first name", 1, null)
+                    .apply(event.getValue(), new ValueContext(firstNameField))
+            if (result.isError()) {
+                UserError error = new UserError(result.getErrorMessage())
+                firstNameField.setComponentError(error)
+            } else {
+                firstNameField.setComponentError(null)
+            }
+        })
+
+        this.lastNameField.addValueChangeListener({ event ->
+            ValidationResult result = new StringLengthValidator("Please input a valid last name",
+                1, null).apply(event.getValue(), new ValueContext(this.lastNameField))
+            if (result.isError()) {
+                UserError error = new UserError(result.getErrorMessage())
+                lastNameField.setComponentError(error)
+            } else {
+                lastNameField.setComponentError(null)
+            }
+        })
+
+        this.emailField.addValueChangeListener({ event ->
+            ValidationResult result = new EmailValidator("Please input a valid email address")
+                .apply(event.getValue(), new ValueContext(emailField))
+            if (result.isError()) {
+                UserError error = new UserError(result.getErrorMessage())
+                emailField.setComponentError(error)
+            } else {
+                emailField.setComponentError(null)
+            }
+        })
+
+        this.submitButton.addClickListener({ event ->
+            String firstName = this.firstNameField.getValue().trim()
+            String lastName = this.lastNameField.getValue().trim()
+            String title = this.titleField.getValue()?.trim() ?: "None"
+            String email = this.emailField.getValue().trim()
+            List<Affiliation> affiliations = new ArrayList()
+            if (affiliationComboBox.selectedItem.isPresent() ) {
+                affiliations.add(affiliationComboBox.getSelectedItem().get())
+            }
+            if (allValuesValid()) {
+                controller.createNewCustomer(firstName, lastName, title, email, affiliations)
+            } else {
+                this.viewModel.failureNotifications.add("Please fill out the customer information correctly.")
+            }
+        })
+    }
 }