+++ CHANGELOG.rst
@@ -5,14 +5,22 @@ Changelog
 This document records all notable changes to `HTTPie <https://httpie.org>`_.
 This project adheres to `Semantic Versioning <https://semver.org/>`_.
 
-
-
-1.0.0
+1.0.0 (2020-09-22)
 ------------------
 
+**Added**
+
+* Implement CreateCustomer use case
+* Create HTML template for offers
 * Create DependencyManager class
 * Maven structure and dependency management of core libraries
 * Create empty infrastructure classes
 * Create usecase classes and interfaces
 * Created the project using QUBE
+
+**Fixed**
+
+**Dependencies**
+
+**Deprecated**